requirements.txt
aio-pika==9.4.3
websockets==12.0
uvloop==0.19.0; platform_system != "Windows"
